Update help.txt
The args= prefix is no longer needed in the examples. Updated the examples to reflect this change.

Added descriptions for the new arguments json / j and gpt / g that are listed in the README.md.

Corrected the weird spacing in lines 21-25.

Added more examples to help users understand how to use the new arguments and the updated format.